Moving filler.insert() helper structs to a separate file (+ templating)
